Update the libraries
Also, I've changed them to use the CDN with cdnjs. I thought this would
be more beneficial for visitors but not the case for site authors since
a couple of security-related features (fingerprinting and subresource
integrity with their own resources) has been stripped. Overall, I do see
this presents more pros than cons. It is easier to maintain both for
site authors and the theme maintainers (READ: me).