Update the integration of Simple Icons
Now, it integrates the icon set through Hugo modules. Unfortunately,
this has the consequence of requiring the Go runtime as one of the
prerequisites for installing the theme.

The 'img' key from the individual link object is also removed because it
is a bit of a burden to maintain it. Debugging this is not nice with all
of the what-ifs and it can be a source of confusion so I'll be removing
it for the time being. I've had not much need for this feature but
contributions are still welcome.